+++ src/Domain/Checker.fs
@@ -7,6 +7,7 @@ module Checker =
     let rec private allTypeNamesOfDefinition = function
         | Type name -> [ name ]
         | Function { Arguments = args; Returns = rets } -> (args |> List.collect allTypeNamesOfDefinition) @ (rets |> allTypeNamesOfDefinition)
+        | Handler { Name = name; Handles = handles } -> name :: (handles |> allTypeNamesOfDefinition)
         | Tuple names -> names |> List.collect allTypeNamesOfDefinition
         | Option name -> name |> allTypeNamesOfDefinition
         | List name -> name |> allTypeNamesOfDefinition