+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Add tuc part
     - `ReadEvent` from stream
 - Show better error message for undefined component field
+- Rename a `only-parse` (`-p`) option to `only-resolved` (`-r`) in `domain:check` command
+- Add domain type for `Handler`
 
 ## 0.1.0 - 2020-08-12
 - Initial implementation